ajout des RestController pour réaliser les opérations CRUD sur les patients

+++ patient/src/main/java/com/mediscreen/patient/service/PatientService.java
@@ -1,12 +1,15 @@
 package com.mediscreen.patient.service;
 
+import com.mediscreen.patient.exception.PatientNotFoundException;
 import com.mediscreen.patient.model.Patient;
 import com.mediscreen.patient.repository.PatientRepository;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
+@Slf4j
 @Service
 public class PatientService {
 
@@ -20,34 +23,53 @@ public class PatientService {
         return patientRepository.findAll();
     }
 
-    public Optional<Patient> getPatient(String lastName, String firstName){
+    public Optional<Patient> getPatient(String lastName, String firstName) {
         return patientRepository.findByLastNameAndFirstName(lastName, firstName);
     }
 
-    public Optional<Patient> getPatientById(String id){
-        return patientRepository.findById(id);
+    public Patient getPatientById(String id) {
+        Optional<Patient> optionalPatient = patientRepository.findById(id);
+        if(optionalPatient.isEmpty()){
+            log.warn("patient with id : " + id + " is not found");
+            throw new PatientNotFoundException("patient with id : " + id + " is not found");
+        } else {
+            log.info("patient is found");
+            return optionalPatient.get();
+        }
     }
 
-    public void addPatient(Patient patient) {
-        patientRepository.insert(patient);
+    public Patient addPatient(Patient patient) {
+        return patientRepository.insert(patient);
     }
 
-    public Patient updatePatient(String id, Patient patient){
-        Optional<Patient> optPatient = patientRepository.findById(id);
-        Patient patientToUpdate = optPatient.get();
-        patientToUpdate.setFirstName(patient.getFirstName());
-        patientToUpdate.setLastName(patient.getLastName());
-        patientToUpdate.setFamily(patient.getFamily());
-        patientToUpdate.setGiven(patient.getGiven());
-        patientToUpdate.setDob(patient.getDob());
-        patientToUpdate.setSex(patient.getSex());
-        patientToUpdate.setAddress(patient.getAddress());
-        patientToUpdate.setPhone(patient.getPhone());
-        patientRepository.save(patientToUpdate);
-        return patientToUpdate;
+    public Patient updatePatient(String id, Patient patient) {
+        Optional<Patient> optionalPatient = patientRepository.findById(id);
+        if (optionalPatient.isEmpty()) {
+            log.warn("patient not found for this id : " + id);
+            throw new PatientNotFoundException("Patient with id " + id + " is not found");
+        } else {
+            Patient patientToUpdate = optionalPatient.get();
+            patientToUpdate.setFirstName(patient.getFirstName());
+            patientToUpdate.setLastName(patient.getLastName());
+            patientToUpdate.setFamily(patient.getFamily());
+            patientToUpdate.setGiven(patient.getGiven());
+            patientToUpdate.setDob(patient.getDob());
+            patientToUpdate.setSex(patient.getSex());
+            patientToUpdate.setAddress(patient.getAddress());
+            patientToUpdate.setPhone(patient.getPhone());
+            patientRepository.save(patientToUpdate);
+            return patientToUpdate;
+        }
     }
 
-    public void deletePatient(String id){
-       patientRepository.deleteById(id);
+    public void deletePatient(String id) {
+        Optional<Patient> optionalPatient = patientRepository.findById(id);
+        if(optionalPatient.isEmpty()){
+            log.warn("patient not found for this id : " + id);
+            throw new PatientNotFoundException("Patient with id " + id + " is not found");
+        } else {
+            patientRepository.deleteById(id);
+            log.info("patient with the id " + id + " has been deleted");
+        }
     }
 }
